requirements.txt
tornado
boto3
google-cloud-storage
google-cloud-speech
Wave
mutagen
requests